% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmixfun.R
\name{dmix}
\alias{dmix}
\title{Probability Density Function (PDF) for Normal Mixture}
\usage{
dmix(x)
}
\arguments{
\item{x}{Numeric value at which to evaluate the PDF.}
}
\value{
The PDF value at `x`.
}
\description{
Computes the PDF of a normal mixture distribution with given parameters.
}
