% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmixfun.R
\name{pmix}
\alias{pmix}
\title{Cumulative Distribution Function (CDF) for Normal Mixture}
\usage{
pmix(q)
}
\arguments{
\item{q}{Numeric value at which to evaluate the CDF.}
}
\value{
The CDF value at `q`.
}
\description{
Computes the CDF of a normal mixture distribution with given parameters.
}
