% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmixfun.R
\name{qmix}
\alias{qmix}
\title{Quantile Function (Inverse CDF) for Normal Mixture}
\usage{
qmix(p)
}
\arguments{
\item{p}{Probability value (between 0 and 1).}
}
\value{
The quantile value corresponding to probability `p`.
}
\description{
Computes the quantile function for a normal mixture distribution.
}
