% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmixfun.R
\name{rmix}
\alias{rmix}
\title{Random Sampling from Normal Mixture}
\usage{
rmix(n)
}
\arguments{
\item{n}{Number of samples to generate.}
}
\value{
A vector of `n` random samples.
}
\description{
Generates random samples from a normal mixture distribution.
}
